requirements-dev.txt
# Development Requirements
# Additional packages for development environment
# Install with: pip install -r requirements-dev.txt

# Include production requirements
-r requirements.txt

# Development Tools
django-debug-toolbar==4.4.6  # Debug toolbar for development
django-extensions==3.2.3     # Useful Django extensions
flake8==7.1.1                # Code linting
black==24.8.0                # Code formatting
isort==5.13.2                # Import sorting

# Testing
pytest==8.3.2               # Testing framework
pytest-django==4.8.0        # Django integration for pytest
coverage==7.6.1             # Code coverage reports
factory-boy==3.3.1          # Test data generation

# Documentation
Sphinx==8.0.2               # Documentation generator
sphinx-rtd-theme==3.0.0     # ReadTheDocs theme

# Production Requirements
# Optimized for production deployment
# Install with: pip install -r requirements-prod.txt

# Include base requirements
-r requirements.txt

# Security & Monitoring
django-cors-headers==4.4.0   # CORS handling
django-csp==3.8              # Content Security Policy
sentry-sdk==2.14.0           # Error monitoring and logging

# Performance & Caching
redis==5.0.8                 # Redis client for caching
django-redis==5.4.0          # Django Redis cache backend
celery==5.3.4                # Task queue for background jobs

# Additional Production Tools
django-storages==1.14.4      # Cloud storage backends (AWS S3, etc.)
boto3==1.35.14               # AWS SDK (if using S3)


